Change int with double

+++ src/main/java/it/unibo/pensilina14/bullet/ballet/graphics/sprite/PhysicalObjectSpriteFactory.java
@@ -4,10 +4,10 @@ import java.io.IOException;
 
 public interface PhysicalObjectSpriteFactory {
 
-    PhysicalObjectSprite generateDynamicObstacleSprite(final int x, final int y) throws IOException;
-    PhysicalObjectSprite generateStaticObstacleSprite(final int x, final int y) throws IOException;
-    PhysicalObjectSprite generateHealingItemSprite(final int x, final int y) throws IOException;
-    PhysicalObjectSprite generateDamagingItemSprite(final int x, final int y) throws IOException;
-    PhysicalObjectSprite generatePoisoningItemSprite(final int x, final int y) throws IOException;
+    PhysicalObjectSprite generateDynamicObstacleSprite(final double x, final double y) throws IOException;
+    PhysicalObjectSprite generateStaticObstacleSprite(final double x, final double y) throws IOException;
+    PhysicalObjectSprite generateHealingItemSprite(final double x, final double y) throws IOException;
+    PhysicalObjectSprite generateDamagingItemSprite(final double x, final double y) throws IOException;
+    PhysicalObjectSprite generatePoisoningItemSprite(final double x, final double y) throws IOException;
     
 }
